+++ eslint.config.js
@@ -40,6 +40,8 @@ module.exports = defineConfig([
   {
     files: ['**/*.html'],
     extends: [angular.configs.templateRecommended, angular.configs.templateAccessibility],
-    rules: {},
+    rules: {
+      '@angular-eslint/template/prefer-static-string-properties': 'error',
+    },
   },
 ]);